requirements.txt
mutagen>=1.47.0
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.1
